Create abstraction for Mock Chart components
