+++ appinventor/appengine/src/com/google/appinventor/client/editor/simple/components/MockLineChart.java
@@ -13,16 +13,8 @@ import org.pepstock.charba.client.resources.ResourcesType;
 
 import java.util.ArrayList;
 
-public final class MockLineChart extends MockVisibleComponent {
-    private static final String PROPERTY_DESCRIPTION = "Description";
-
+public final class MockLineChart extends MockChart<LineChart> {
     public static final String TYPE = "LineChart";
-
-    private LineChart lineChartWidget;
-
-    static {
-        ResourcesType.setClientBundle(EmbeddedResources.INSTANCE);
-    }
 
     /**
      * Creates a new instance of a Mock Line Chart component.
@@ -33,15 +25,13 @@ public final class MockLineChart extends MockVisibleComponent {
         super(editor, TYPE, images.image());
 
         // Initialize Line Chart mock Widget
-        lineChartWidget = new LineChart();
+        chartWidget = new LineChart();
 
+        // Load test data set
         loadTestData();
 
-        // Chart widget setup
-        lineChartWidget.getOptions().setMaintainAspectRatio(false);
-        lineChartWidget.getOptions().getTitle().setDisplay(true);
-        lineChartWidget.setStylePrimaryName("ode-SimpleMockComponent");
-        initComponent(lineChartWidget);
+        // Initialize the Chart
+        initChart();
     }
 
     /**
@@ -49,7 +39,7 @@ public final class MockLineChart extends MockVisibleComponent {
      */
     private void loadTestData() {
         // Initialize data set
-        LineDataset dataset = lineChartWidget.newDataset();
+        LineDataset dataset = chartWidget.newDataset();
 
         // Construct test data
         dataset.setData(1, 7, 5, 4);
@@ -63,45 +53,14 @@ public final class MockLineChart extends MockVisibleComponent {
         dataset.setLabel("Data");
 
         // Set the data set to the chart
-        lineChartWidget.getData().setDatasets(dataset);
+        chartWidget.getData().setDatasets(dataset);
 
         // Set x value labels
-        lineChartWidget.getData().setLabels("1", "2", "3", "4");
-    }
-
-    private void setDescriptionProperty(String text) {
-        lineChartWidget.getOptions().getTitle().setText(text);
-    }
-
-    /*
-     * Sets the Line Chart's BackgroundColor property to a new value.
-     */
-    private void setBackgroundColorProperty(String text) {
-        if (MockComponentsUtil.isDefaultColor(text)) {
-            text = "&HFFFFFFFF";  // white
-        }
-        MockComponentsUtil.setWidgetBackgroundColor(lineChartWidget, text);
-    }
-
-    @Override
-    public int getPreferredWidth() {
-        return ComponentConstants.VIDEOPLAYER_PREFERRED_WIDTH;
-    }
-
-    @Override
-    public int getPreferredHeight() {
-        return ComponentConstants.VIDEOPLAYER_PREFERRED_HEIGHT;
+        chartWidget.getData().setLabels("1", "2", "3", "4");
     }
 
     @Override
     public void onPropertyChange(String propertyName, String newValue) {
         super.onPropertyChange(propertyName, newValue);
-
-        if (propertyName.equals(PROPERTY_DESCRIPTION)) {
-            setDescriptionProperty(newValue);
-            lineChartWidget.draw();
-        } else if (propertyName.equals(PROPERTY_NAME_BACKGROUNDCOLOR)) {
-            setBackgroundColorProperty(newValue);
-        }
     }
 }
